Qt text search crash fix
About 5 weeks ago, commit 11995764ed15cb87eccefd1e1a251f5e07211486
changed Qt's searching in text windows to be able to find a match
on the very first line.  It assumed that the first line would be
the current line except when repeating the same search after a
match.  But after a failed search the current line index is -1 and
starting a new search would crash trying to look that line up.