fixes37.0 entry about engraving
Revert part of commit 171873aea5b95e69436488c39b8267a97ac3e06e
that moved a fixes entry from the 'fixes to 3.6.x' section
into the 'post-3.6.x, exposed by git' section.  It wasn't due
to the conversion of engraving from multi-turn action into an
occupation, it was fixing a bug that had been around for 5 or 6
years.  Moot now; the conversion to occupation removed it.